+++ gen/bobgen-psql/driver/parser/walker.go
@@ -189,6 +189,12 @@ func (w *walker) walk(a any) nodeInfo {
 	case *pg.InsertStmt:
 		info = w.walkInsertStmt(a)
 
+	case *pg.UpdateStmt:
+		info = w.walkUpdateStmt(a)
+
+	case *pg.DeleteStmt:
+		info = w.walkDeleteStmt(a)
+
 	case *pg.ParamRef:
 		info = w.walkParamRef(a)
 
@@ -399,6 +405,28 @@ func (w *walker) walkInsertStmt(a *pg.InsertStmt) nodeInfo {
 				}
 			}
 		}
+	}
+
+	return info
+}
+
+func (w *walker) walkUpdateStmt(a *pg.UpdateStmt) nodeInfo {
+	info := w.reflectWalk(reflect.ValueOf(a))
+	info.start = w.getStartOfTokenBefore(info.start, pg.Token_UPDATE)
+
+	if err := verifyUpdateStatement(a, info); err != nil {
+		w.errors = append(w.errors, err)
+	}
+
+	return info
+}
+
+func (w *walker) walkDeleteStmt(a *pg.DeleteStmt) nodeInfo {
+	info := w.reflectWalk(reflect.ValueOf(a))
+	info.start = w.getStartOfTokenBefore(info.start, pg.Token_DELETE_P)
+
+	if err := verifyDeleteStatement(a, info); err != nil {
+		w.errors = append(w.errors, err)
 	}
 
 	return info